backend/requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.38.0
sqlalchemy==2.0.44
beautifulsoup4==4.14.2
requests==2.32.5
pydantic==2.12.4
langchain-core==1.0.3
langchain-community==0.4.1
langchain-google-genai==3.0.1
python-dotenv==1.2.1
pymysql==1.1.1
psycopg2-binary==2.9.9